+++ SeleniumLevel3/src/test/java/datatype/VietJet/Booking.java
@@ -2,7 +2,7 @@ package datatype.VietJet;
 
 import java.util.Date;
 
-import utils.helper.DataHelper;
+import utils.helper.ResourceHelper;
 import utils.helper.DateTimeHelper;
 
 public class Booking {
@@ -10,7 +10,7 @@ public class Booking {
 	private String flightClass = null;
 	private String departureFrom = null;
 	private String departureTo = null;
-	private String departureDate = DateTimeHelper.getDateString(new Date(), DataHelper.SHARED_DATA.get().date_format);
+	private String departureDate = DateTimeHelper.getDateString(new Date(), ResourceHelper.SHARED_DATA.get().date_format);
 	private String departureTime = null;
 	private double departureFare = 0;
 	private double departureCharge = 0;
@@ -18,7 +18,7 @@ public class Booking {
 	private double departureTotal = 0;
 	private String returnFrom = departureTo;
 	private String returnTo = departureFrom;
-	private String returnDate = DateTimeHelper.getDateString(DateTimeHelper.plusDays(1), DataHelper.SHARED_DATA.get().date_format);
+	private String returnDate = DateTimeHelper.getDateString(DateTimeHelper.plusDays(1), ResourceHelper.SHARED_DATA.get().date_format);
 	private String returnTime = null;
 	private double returnFare = 0;
 	private double returnCharge = 0;
@@ -73,7 +73,7 @@ public class Booking {
 	}
 
 	public void setDepartureDate(Date departureDate) {
-		this.departureDate = DateTimeHelper.getDateString(departureDate, DataHelper.SHARED_DATA.get().date_format);
+		this.departureDate = DateTimeHelper.getDateString(departureDate, ResourceHelper.SHARED_DATA.get().date_format);
 	}
 
 	public String getDepartureTime() {
@@ -141,7 +141,7 @@ public class Booking {
 	}
 
 	public void setReturnDate(Date returnDate) {
-		this.returnDate = DateTimeHelper.getDateString(returnDate, DataHelper.SHARED_DATA.get().date_format);
+		this.returnDate = DateTimeHelper.getDateString(returnDate, ResourceHelper.SHARED_DATA.get().date_format);
 	}
 
 	public String getReturnTime() {
@@ -243,9 +243,9 @@ public class Booking {
 	// Update all the values if need to calculate after providing data.
 	public Booking compileData() {
 		if (this.departureDate.contains(":"))
-			setDepartureDate(DateTimeHelper.getDateString(DateTimeHelper.mapDate(this.departureDate), DataHelper.SHARED_DATA.get().date_format));
+			setDepartureDate(DateTimeHelper.getDateString(DateTimeHelper.mapDate(this.departureDate), ResourceHelper.SHARED_DATA.get().date_format));
 		if (this.returnDate.contains(":"))
-			setReturnDate(DateTimeHelper.getDateString(DateTimeHelper.mapDate(this.returnDate), DataHelper.SHARED_DATA.get().date_format));
+			setReturnDate(DateTimeHelper.getDateString(DateTimeHelper.mapDate(this.returnDate), ResourceHelper.SHARED_DATA.get().date_format));
 		setReturnFrom(getDepartureTo());
 		setReturnTo(getDepartureFrom());
 		return this;
@@ -262,86 +262,5 @@ public class Booking {
 				+ ", returnTax=" + returnTax + ", returnTotal=" + returnTotal + ", grandTotal=" + grandTotal
 				+ ", numberOfAdults=" + numberOfAdults + ", numberOfChildren=" + numberOfChildren + ", numberOfInfants="
 				+ numberOfInfants + "]";
-	}
-
-	public enum FlightClass {
-		ECO("Eco", "Eco"), PROMO("Promo", "Promo"), SKYBOSS("", "Skyboss"), NONE("", "");
-
-		private String key;
-		private String value;
-
-		FlightClass(String key, String value) {
-			this.key = key;
-			this.value = value;
-		}
-
-		public String getValue() {
-			return this.value;
-		}
-
-		public String getKey() {
-			return this.key;
-		}
-	}
-
-	public enum FlightOption {
-		ONEWAY("One Way"), RETURN("Return"), NONE("");
-		private String option;
-
-		FlightOption(String option) {
-			this.option = option;
-		}
-
-		public String getValue() {
-			return this.option;
-		}
-	}
-
-	public enum LocationOption {
-		SGN("SGN", "Ho Chi Minh"), HAN("HAN", "Ha Noi"), BMV("BMV", "Buon Ma Thuot"), VCA("VCA", "Can Tho"),
-		VCL("VCL", "Chu Lai"), DLI("DLI", "Da Lat"), DAD("DAD", "Da Nang"), VDH("VDH", "Dong Hoi"),
-		HPH("HPH", "Hai Phong"), HUI("HUI", "Hue"), CXR("CXR", "Nha Trang"), PQC("PQC", "Phu Quoc"),
-		PXU("PXU", "Pleiku"), UIH("UIH", "Quy Nhon"), THD("THD", "Thanh Hoa"), TBB("TBB", "Tuy Hoa"),
-		VDO("VDO", "Van Don"), VII("VII", "Vinh"), BKK("BKK", "Bangkok - Suvarnabhumi"), CNX("CNX", "Chiang Mai"),
-		CEI("CEI", "Chiang Rai"), KBV("KBV", "Krabi"), HKT("HKT", "Phuket"), UTH("UTH", "Udon Thani"),
-		PUS("PUS", "Busan"), TAE("TAE", "Daegu"), HAK("HAK", "Haikou"), HND("HND", "Haneda Tokyo"),
-		HFE("HFE", "Hefei Xinqiao"), HKG("HKG", "Hong Kong – Terminal 1"), KHH("KHH", "Kaohsiung"),
-		KUL("KUL", "Kuala Lumpur – KLIA"), MFM("MFM", "Macau"), BOM("BOM", "Mumbai - Terminal 2"),
-		DEL("DEL", "New Delhi"), DPS("DPS", "Ngurah Rai - Bali"), KIX("KIX", "Osaka - Terminal 1"),
-		PNH("PNH", "Phnom Penh"), ICN("ICN", "Seoul"), REP("REP", "Siem Reap"), SIN("SIN", "Singapore - Terminal 4"),
-		RMQ("RMQ", "Taichung"), TNN("TNN", "Tainan"), TPE("TPE", "Taipei - Terminal 1"), NRT("NRT", "Tokyo Narita"),
-		RGN("RGN", "Yangon"), NONE("", "");
-
-		private String key;
-		private String value;
-
-		LocationOption(String key, String value) {
-			this.key = key;
-			this.value = value;
-		}
-
-		public String getValue() {
-			return this.value;
-		}
-
-		public String getKey() {
-			return this.key;
-		}
-
-		public static String getCode(String value) {
-			for (LocationOption e : LocationOption.values()) {
-				if (e.value.equals(value))
-					return e.key;
-			}
-			return null;
-		}
-
-		public static String getValue(String code) {
-			for (LocationOption e : LocationOption.values()) {
-				if (e.key.equals(code))
-					return e.value;
-			}
-			return null;
-		}
 	}
 }
