+++ SeleniumLevel3/src/test/java/pages/VietJet/HomePage.java
@@ -8,10 +8,10 @@ import element.wrapper.web.RadioButton;
 import element.wrapper.web.TextBox;
 import helper.LocatorHelper;
 import utils.constant.Constants;
-import utils.helper.DataHelper;
+import utils.helper.ResourceHelper;
 
 public class HomePage {
-	LocatorHelper locator = new LocatorHelper(Constants.LOCATOR_FOLDER_PATH + DataHelper.SHARED_DATA.get().appName, getClass().getSimpleName());
+	LocatorHelper locator = new LocatorHelper(Constants.LOCATOR_FOLDER_PATH + ResourceHelper.SHARED_DATA.get().appName, getClass().getSimpleName());
 	// Elements
 
 	protected DropDown cbxLanguage = new DropDown(locator.getLocator("cbxLanguage"));
