+++ SeleniumLevel3/src/test/java/tests/VietJet/TS_VietJet.java
@@ -11,7 +11,7 @@ import pages.VietJet.SelectTravelOptionsPage;
 import tests.TestBase;
 import utils.assertion.SoftAssertion;
 import utils.constant.Constants;
-import utils.helper.DaoHelper;
+import utils.helper.DataHelper;
 import utils.helper.Logger;
 
 public class TS_VietJet extends TestBase {
@@ -22,7 +22,7 @@ public class TS_VietJet extends TestBase {
 		SoftAssertion softAssert = new SoftAssertion();
 		
 		Logger.info("Precondition: Initial Data");
-		DaoHelper dataHelper = new DaoHelper(Constants.DATA_FOLDER + this.appName, "TC01");
+		DataHelper dataHelper = new DataHelper(Constants.DATA_FOLDER + this.appName, "TC01");
 		HomePage homePage = new HomePage();
 		Booking booking = dataHelper.getDataObject(Booking.class).compileData();
 		Logger.info(booking.toString());
@@ -73,7 +73,7 @@ public class TS_VietJet extends TestBase {
 		
 		SoftAssertion softAssert = new SoftAssertion();
 		
-		DaoHelper dataHelper = new DaoHelper(Constants.DATA_FOLDER + "VietJet", new Object() {
+		DataHelper dataHelper = new DataHelper(Constants.DATA_FOLDER + "VietJet", new Object() {
 		}.getClass().getEnclosingMethod().getName());
 		
 		Logger.info("1. Navigate to https://www.vietjetair.com/Sites/Web/vi-VN/Home.");
